Feature/SK-718 | Adds option to install postgres cluster via CloudnativePG plus other clean up (#110)
A few changes:
If the Cloudnative PG operator is installed on the cluster, then either the chart can install a postgres cluster, or if the user has already installed a cluster, then Studio can connect to it. The only settings that you need to change are studio.cloudnativepg.enabled and studio.cloudnativepg.create_new, no need to set passwords for Postgres.
If a Minio cluster is installed via the Minio operator, and named "minio-fedn", then no changes to the values file are required, and the user no longer needs to then create a secret manually (assuming that the name of the user secret is the default one set by Minio and not changed by the user).
If postgres and the Minio cluster is set up as above then backups will be stored in a bucket in that Minio instance. Can be configured to be a different Minio store.
Removed mediavol.
Added environment variables to Studio deployment that are needed for doing log aggregation of FEDn logs in Studio.